code readability improvements
